# cmd/senda/docs/swagger.yaml
basePath: /
definitions:
  cmd_senda.DocErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      request_id:
        type: string
    type: object
  cmd_senda.DocErrorResponse:
    properties:
      error:
        $ref: '#/definitions/cmd_senda.DocErrorDetail'
    type: object
  cmd_senda.DocGenericObject:
    type: object
...
info:
  contact: {}
  description: Multi-tenant email orchestration API for Senda.
  title: Senda API
  version: "1.0"
paths:
...
  /api/v1/webhooks/ses/inbound:
    post:
      description: Auto-generated route stub for OpenAPI + MCP discovery.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_senda.DocGenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd_senda.DocErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_senda.DocErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/cmd_senda.DocErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd_senda.DocErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/cmd_senda.DocErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/cmd_senda.DocErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cmd_senda.DocErrorResponse'
      summary: POST /api/v1/webhooks/ses/inbound
      tags:
      - webhooks
...
securityDefinitions:
  ManagementBearer:
    description: 'OIDC bearer token sent as `Authorization: Bearer <jwt>`.'
    in: header
    name: Authorization
    type: apiKey
  WorkspaceAPIKeyBearer:
    description: 'Workspace API key sent as `Authorization: Bearer senda_live_...`.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"